Restructure test directories and improve temp dir error handling
