+++ .gitignore
@@ -19,6 +19,6 @@ main
 .vscode/
 
 # File created running tests
-tests/tmp/
+tests/**/tmp/
 
 .DS_Store
